requirements.txt
fastapi
uvicorn
Pillow
tensorflow
nump
python-multipart
